requirements.txt
pyodbc>=5.0.0
pandas>=2.0.0
pymysql>=1.1.0
SQLAlchemy>=2.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
